requirements.txt
django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
cryptography>=42.0
eth-account>=0.11
web3>=6.0
python-decouple>=3.8
reportlab>=4.2
Pillow>=10.0
qrcode[pil]>=7.4
